requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
backoff==2.2.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==14.0
croniter==2.0.5
dagit==1.7.12
dagster==1.7.12
dagster-graphql==1.7.12
dagster-pipes==1.7.12
dagster-webserver==1.7.12
db-dtypes==1.2.0
docstring_parser==0.16
exceptiongroup==1.2.1
filelock==3.15.4
fsspec==2024.6.1
google-api-core==2.19.1
google-auth==2.31.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.62.2
grpcio-status==1.64.1
gspread==6.1.2
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pendulum==3.0.0
proto-plus==1.24.0
protobuf==5.27.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
structlog==24.2.0
tabulate==0.9.0
time-machine==2.14.2
tomli==2.0.1
toposort==1.10
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.2
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
